pyproject.toml
[project]
name = "langchain-mcp-adaptors"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.1.7",
    "python-dotenv>=1.1.0",
]
